move releaseIfNeeded to base class
